Add files via upload
Added random placement functionality to Battleship_Game. Designed Battle_Ship game so that Oceanboard is not needed. Removed Ship class functionality in favor of integers in the HPlayer class. Player abstract class omitted as for the moment we are assuming two human players.

All functions were working as-intended on linprog as of 11/27. If you copy the Battleship_game and HPlayer classes, the game should be able to work as-intended in the terminal, which has been the following:

1. Prior to the game starting, the ship-boards are displayed on-screen, the first board is player 1's ships and second is player 2's ships.
2. Player 1 will make their move by choosing a row and column which they want to fire on (error checking included of course). This is received by user keyboard input on the terminal
3. Once a player moves, their "radar" is displayed on screen. a message should also appear on the terminal with the user has either hit or missed.
4. Player 2 plays with exact same aforementioned mechanics

Once both players have played, their total remaining health (the sum of all of their ships health) is displayed on the terminal. The loop repeats until the game condition (one of the two players have 0 health) is reached.

If you have any further questions please let me know.